+++ site/svelte.config.js
@@ -4,9 +4,9 @@ import adapter from '@sveltejs/adapter-static';
 const config = {
   kit: {
     adapter: adapter({
-      fallback: '404.html'
-    })
-  }
+      fallback: '404.html',
+    }),
+  },
 };
 
 export default config;