+++ src/plugin/index.ts
@@ -1,5 +1,5 @@
 export { norgPlugin, type ArboriumConfig, type NorgPluginOptions } from './plugin';
-export type { NorgParseResult, TocEntry, InlineComponent } from '@parser';
+export type { NorgParseResult, TocEntry, EmbedComponent } from '@parser';
 export type { HtmlModule } from './types/html';
 export type { SvelteModule } from './types/svelte';
 export type { ReactModule } from './types/react';